requirements.txt
sqlmodel
fastapi[standard]
pymysql
uvicorn
sqlalchemy
python-multipart
azure-storage-blob
python-dotenv
cryptography